Fix LayerScale dtype under autocast
self.gamma is fp32 (default Parameter dtype). Under autocast(bf16), x
is bf16 coming into the LayerScale call. The multiply 'gamma * x'
follows PyTorch's normal type promotion rule (fp32·bf16 → fp32), so
LayerScale silently upcasts the residual branch back to fp32. That
fp32 output then lands in the bf16 x_after_group tensor and blows up
at the indexed assignment.

Cast gamma to x.dtype before the multiply so the residual stays in
whatever dtype autocast is using.

+++ src/model/encoder.py
@@ -64,7 +64,9 @@ class LayerScale(nn.Module):
         self.gamma = nn.Parameter(init_value * torch.ones(dim))
 
     def forward(self, x: Tensor) -> Tensor:
-        return self.gamma * x
+        # Cast gamma to x's dtype so autocast (bf16/fp16) doesn't upcast
+        # the residual branch to fp32 at the multiply.
+        return self.gamma.to(x.dtype) * x
 
 
 class WindowedAttention(nn.Module):
